requirements.txt
selenium==3.141.0
pytest==6.2.0
allure-pytest==2.9.42
